# contracts/finance.openapi.yaml
openapi: 3.0.3
info:
  title: erp-finance · 财务管理
  version: 1.0.0
  description: |
    对外 REST 面（组件之间走 gRPC）。不暴露 CheckPeriodOpen 与 BatchGetCreditExposure：
    前者是组件之间期间锁的咨询入口，后者是给别的组件的 gRPC 客户端一次取多个客户用的。
    每条路径都要求登录；可见的数据按调用者的法人访问授权（legal_entity_access）过滤。

servers:
  - url: /erp/finance

# /healthz 不在这份契约里：它由 be-sdk-go 的 NewGinEngine 统一挂在裸根路径 /healthz，
# 只报告进程活着。

paths:
  /periods/{period}/close:
    post:
      summary: ClosePeriod（关账，可逆）
      operationId: closeFinancePeriod
      parameters:
        - { name: period, in: path, required: true, schema: { type: string }, description: 格式 YYYY-MM }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PeriodOpRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PeriodOpResponse" }

  /periods/{period}/reopen:
    post:
      summary: ReopenPeriod（反关账）
      operationId: reopenFinancePeriod
      parameters:
        - { name: period, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PeriodOpRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PeriodOpResponse" }

  /periods/{period}/lock:
    post:
      summary: LockPeriod（锁定，不可逆）
      operationId: lockFinancePeriod
      parameters:
        - { name: period, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PeriodOpRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PeriodOpResponse" }

  /credit-exposure/{customer_id}:
    get:
      summary: GetCreditExposure
      operationId: getCreditExposure
      parameters:
        - { name: customer_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreditExposure" }

  /entries:
    get:
      summary: ListEntries
      operationId: listFinanceEntries
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
          description: 游标分页，没有 offset 参数——深分页在契约层面不可表达
        - name: page_size
          in: query
          schema: { type: integer }
        - name: period
          in: query
          schema: { type: string }
        - name: status_filter
          in: query
          schema: { $ref: "#/components/schemas/EntryStatus" }
        - name: source_doc_id
          in: query
          schema: { type: string }
          description: 只看这张源单生成的凭证（与 JournalEntry.source_doc_id 相等），例如订单详情页反查对应凭证
        - name: source_doc_type
          in: query
          schema: { type: string }
          description: 只看这类源单生成的凭证，例如 order（销售订单）、movement（库存流水）、reversal（红字冲销）
        - name: created_after
          in: query
          schema: { type: string, format: date-time }
        - name: created_before
          in: query
          schema: { type: string, format: date-time }
          description: 不传时间范围时默认最近 90 天
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items: { $ref: "#/components/schemas/JournalEntry" }
                  next_cursor: { type: string }
    post:
      summary: PostManualEntry（手工凭证，自动凭证走事件消费不走这里）
      operationId: postManualFinanceEntry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostManualEntryRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JournalEntry" }

  /entries/{id}:
    get:
      summary: GetEntry
      operationId: getFinanceEntry
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JournalEntry" }

  /entries/{id}/reverse:
    post:
      summary: ReverseEntry（红字冲销，不是删除；返回新产生的反向凭证）
      operationId: reverseFinanceEntry
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReverseEntryRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JournalEntry" }

  /ar-ledger:
    get:
      summary: ListARLedger
      operationId: listARLedger
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: page_size
          in: query
          schema: { type: integer }
        - name: customer_id
          in: query
          schema: { type: string }
        - name: created_after
          in: query
          schema: { type: string, format: date-time }
        - name: created_before
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items: { $ref: "#/components/schemas/ARLedgerEntry" }
                  next_cursor: { type: string }

  /ar-ledger/summary:
    get:
      summary: 应收统计（仪表盘）：合计、已核销、未核销与账龄分桶
      description: |
        只统计调用者有法人访问授权（legal_entity_access）的应收。不套列表的默认时间窗口：
        这是对全部应收的统计。账龄 = 统计当天（as_of，UTC）− 到期日（天），按未核销余额
        分桶；还没到期的算在 d0_30。权限键 erp.finance.view。
      operationId: summarizeARLedger
      parameters:
        - name: customer_id
          in: query
          schema: { type: string }
          description: 只统计这个客户
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ARLedgerSummary" }

  /legal-entity-access/{sub}:
    get:
      summary: 列出某个用户能访问的法人（权限键 erp.finance.manage_access）
      operationId: listLegalEntityAccess
      parameters:
        - { name: sub, in: path, required: true, schema: { type: string }, description: 用户在 IAM 里的 sub }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  legal_entity_ids:
                    type: array
                    items: { type: string }
    post:
      summary: 授予某个用户访问一个法人的账（幂等；权限键 erp.finance.manage_access）
      operationId: grantLegalEntityAccess
      parameters:
        - { name: sub, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [legal_entity_id]
              properties:
                legal_entity_id: { type: string }
      responses:
        "200": { description: OK }

  /legal-entity-access/{sub}/{legal_entity_id}:
    delete:
      summary: 撤销某个用户对一个法人的访问（幂等；权限键 erp.finance.manage_access）
      operationId: revokeLegalEntityAccess
      parameters:
        - { name: sub, in: path, required: true, schema: { type: string } }
        - { name: legal_entity_id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }

components:
  schemas:
    PeriodStatus:
      type: string
      enum: [OPEN, CLOSED, LOCKED]

    EntryStatus:
      type: string
      enum: [DRAFT, POSTED]

    PeriodOpRequest:
      type: object
      required: [idempotency_key, legal_entity_id]
      properties:
        idempotency_key: { type: string }
        legal_entity_id: { type: string }

    PeriodOpResponse:
      type: object
      properties:
        status: { $ref: "#/components/schemas/PeriodStatus" }

    CreditExposure:
      type: object
      properties:
        customer_id: { type: string }
        exposure:
          type: string
          description: 金额一律用 string 传 decimal，绝不用 double（跨语言精度）
        version: { type: integer, format: int64 }

    JournalEntryLine:
      type: object
      required: [account_id, debit, credit]
      properties:
        account_id: { type: string }
        debit: { type: string }
        credit: { type: string }
        memo: { type: string }

    JournalEntry:
      type: object
      properties:
        id: { type: string }
        entry_no: { type: string }
        post_no: { type: string, description: 空字符串表示还未过账 }
        period: { type: string }
        legal_entity_id: { type: string }
        status: { $ref: "#/components/schemas/EntryStatus" }
        source_component: { type: string }
        source_doc_type: { type: string }
        source_doc_id: { type: string }
        source_revision: { type: integer, format: int64 }
        lines:
          type: array
          items: { $ref: "#/components/schemas/JournalEntryLine" }
        memo: { type: string }
        version: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
        posted_at: { type: string, format: date-time }

    PostManualEntryRequest:
      type: object
      required: [idempotency_key, legal_entity_id, lines]
      properties:
        idempotency_key:
          type: string
          description: 跨重试幂等键，同一个键只执行一次
        legal_entity_id: { type: string }
        lines:
          type: array
          items: { $ref: "#/components/schemas/JournalEntryLine" }
        memo: { type: string }

    ReverseEntryRequest:
      type: object
      required: [idempotency_key]
      properties:
        idempotency_key: { type: string }
        reason: { type: string }

    ARLedgerEntry:
      type: object
      properties:
        id: { type: string }
        customer_id: { type: string }
        entry_id: { type: string }
        amount: { type: string }
        reconciled_amount: { type: string }
        created_at: { type: string, format: date-time }
        customer_name:
          type: string
          description: 客户名，取自本组件消费 mdm.customer.* 事件维护的客户摘要副本；还没收到过这个客户的事件时为空串
        outstanding:
          type: string
          description: 未核销余额 = amount - reconciled_amount，服务端按十进制算（decimal-as-string）
        due_date:
          type: string
          format: date
          description: 到期日（YYYY-MM-DD），账龄从它算起；上游不带付款条件时等于记账当天

    ARLedgerSummary:
      type: object
      description: 金额一律是恰好两位小数的十进制字符串；没有数据时是 "0.00"
      properties:
        as_of: { type: string, format: date, description: 统计当天（UTC） }
        total_receivable: { type: string, description: 应收合计（sum amount） }
        total_reconciled: { type: string, description: 已核销合计（sum reconciled_amount） }
        outstanding: { type: string, description: 未核销合计 = total_receivable - total_reconciled }
        aging:
          type: object
          description: 未核销余额按账龄（天）分桶，四桶之和等于 outstanding
          properties:
            d0_30: { type: string, description: 0–30 天，含未到期 }
            d31_60: { type: string }
            d61_90: { type: string }
            d90_plus: { type: string, description: 超过 90 天 }